+++ lib/nghttp3_stream.c
@@ -959,9 +959,9 @@ static int stream_pop_outq_entry(nghttp3_stream *stream,
   return 0;
 }
 
-int nghttp3_stream_add_ack_offset(nghttp3_stream *stream, size_t n) {
+int nghttp3_stream_add_ack_offset(nghttp3_stream *stream, uint64_t n) {
   nghttp3_ringbuf *outq = &stream->outq;
-  size_t offset = stream->ack_offset + n;
+  uint64_t offset = stream->ack_offset + n;
   size_t buflen;
   size_t npopped = 0;
   size_t nack;
@@ -973,7 +973,7 @@ int nghttp3_stream_add_ack_offset(nghttp3_stream *stream, size_t n) {
     buflen = nghttp3_buf_len(&tbuf->buf);
 
     if (tbuf->type == NGHTTP3_BUF_TYPE_ALIEN) {
-      nack = nghttp3_min(offset, buflen) - stream->ack_done;
+      nack = (size_t)nghttp3_min(offset, (uint64_t)buflen) - stream->ack_done;
       if (stream->callbacks.acked_data) {
         rv = stream->callbacks.acked_data(stream, stream->node.nid.id, nack,
                                           stream->user_data);
