+++ lib/nghttp3_stream.h
@@ -235,7 +235,7 @@ struct nghttp3_stream {
   size_t outq_offset;
   /* ack_offset is offset acknowledged by peer relative to the first
      element in outq. */
-  size_t ack_offset;
+  uint64_t ack_offset;
   /* ack_done is the number of bytes notified to an application that
      they are acknowledged inside the first outq element if it is of
      type NGHTTP3_BUF_TYPE_ALIEN. */
@@ -343,7 +343,7 @@ int nghttp3_stream_add_outq_offset(nghttp3_stream *stream, size_t n);
  */
 int nghttp3_stream_outq_write_done(nghttp3_stream *stream);
 
-int nghttp3_stream_add_ack_offset(nghttp3_stream *stream, size_t n);
+int nghttp3_stream_add_ack_offset(nghttp3_stream *stream, uint64_t n);
 
 /*
  * nghttp3_stream_is_active returns nonzero if |stream| is active.  In
